refactor(app): extract 5 helper modules from message-timeline
Slice 11b.1 Phase 3b first pass. Pull self-contained concerns out of
the 1211-line message-timeline.tsx into dedicated modules per design
doc §3b.

New modules:
  message-timeline-staging.ts        (103 lines)
    createTimelineStaging — defer-mounts old turns to keep first paint
    cheap on session switch.

  message-timeline-scroll.ts         (101 lines)
    boundaryTarget / boundaryGesture / markBoundaryGesture +
    scrollViewMetricsToTimelineMetrics / scrollViewIntentToTimelineIntent /
    shouldMarkLegacyScrollIntent — pure helpers that translate
    ScrollView intents into the timeline scroll controller's shape.

  message-timeline-comments.tsx      (91 lines)
    MessageComment type + messageComments parser + MessageCommentList
    component — the file-chip rail rendered above each turn for
    synthetic comment parts.

  message-timeline-history-load.tsx  (37 lines)
    LoadEarlierButton — the ghost button at the top of the timeline
    when there are turns older than the current window.

  message-timeline-turn-changes.tsx  (332 lines)
    createTurnChangeFetcher — owns the turnChanges store + imperative
    undo/redo with conflict dialog + auto-prefetch effect + 500ms
    retry timer for transient empty responses + lifecycle cleanup.

message-timeline.tsx: 1211 → 710 lines. Behaviour preserved — module
boundaries follow existing function boundaries; no logic changes.

tsgo --noEmit GREEN (app + ui). bun test (app): 1050 pass / 0 fail.

Followup commit will trim orphaned title-editor / session-export code
left behind by b49b67999 ("drop duplicate session title bar"); that
cleanup is independent of this extraction and reviewable on its own.
